+++ zcash_primitives/src/lib.rs
@@ -15,7 +15,6 @@ pub mod legacy;
 pub mod memo;
 pub mod merkle_tree;
 pub mod note_encryption;
-pub mod pedersen_hash;
 pub mod primitives;
 pub mod sapling;
 pub mod serialize;
