+++ zcash_primitives/src/test_vectors/pedersen_hash_vectors.rs
@@ -1,6 +1,6 @@
 //! Test vectors from https://github.com/zcash-hackworks/zcash-test-vectors/blob/master/sapling_pedersen.py
 
-use crate::pedersen_hash::{test::TestVector, Personalization};
+use crate::sapling::pedersen_hash::{test::TestVector, Personalization};
 
 pub fn get_vectors<'a>() -> Vec<TestVector<'a>> {
     return vec![
